turned weights back to float. There is a problem with the pruning code that needs to be solved

+++ CE_Louvain_V5/community.cpp
@@ -102,7 +102,7 @@ Community::init_partition(char * filename) {
       unsigned int i=0;
       for ( i=0 ; i<neigh_last ; i++) {
 	unsigned int best_comm     = neigh_pos[i];
-	double best_nblinks  = neigh_weight[neigh_pos[i]];
+	float best_nblinks  = neigh_weight[neigh_pos[i]];
 	if (best_comm==comm) {
 	  insert(node, best_comm, best_nblinks);
 	  break;
@@ -141,7 +141,7 @@ Community::neigh_comm(unsigned int node) {
     neigh_weight[neigh_pos[i]]=-1;
   neigh_last=0;
 
-  pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+  pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
 
   unsigned int deg = g.nb_neighbors(node);
 
@@ -181,7 +181,7 @@ Community::miniBatch_assign(long W){
         	    diff = deg - max_deg;
          	    max_deg = deg;
       		}
-		long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(double)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
+		long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(float)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
       		if(sum + tmp + diff *(long)(sizeof(double)+sizeof(unsigned int)) >= W)
         	   break;
       		sum += tmp + diff * (sizeof(double)+sizeof(unsigned int));
@@ -254,7 +254,7 @@ Community::update_sum(int curr, int* max_deg, long sum){
          diff = deg - *max_deg;
          *max_deg = deg; 
        }
-       long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(double)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
+       long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(float)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
        sum += tmp + diff * (sizeof(double)+sizeof(unsigned int));
        return sum; 
 }
@@ -339,7 +339,7 @@ Community::ABFS_preorder(vector<int>& retorder, long W){
 
            //insert the neighbors and remove curr from the queue
            int deg = g.nb_neighbors(curr);
-           pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(curr);
+           pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(curr);
            vector<int> children(deg,-1);
           
            int ch_num = 0; 
@@ -435,7 +435,7 @@ Community::transform(long W){
       int node = order[index];
       int deg = g.nb_neighbors(node);
       g2.degrees[index] = (index==0)?deg:g2.degrees[index-1]+deg;
-      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
       for(int i=0; i<deg; i++){
 	  int neigh = *(p.first+i);
 	  g2.links[link_index] = reverse_map[neigh];
@@ -486,7 +486,7 @@ Community::partition2graph() {
 
 
   for (int i=0 ; i<size ; i++) {
-    pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(i);
+    pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(i);
 
     int deg = g.nb_neighbors(i);
     for (int j=0 ; j<deg ; j++) {
@@ -539,12 +539,12 @@ Community::partition2graph_binary() {
 
   int comm_deg = comm_nodes.size();
   for (int comm=0 ; comm<comm_deg ; comm++) {
-    map<int,double> m;
-    map<int,double>::iterator it;
+    map<int,float> m;
+    map<int,float>::iterator it;
 
     int comm_size = comm_nodes[comm].size();
     for (int node=0 ; node<comm_size ; node++) {
-      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(comm_nodes[comm][node]);
+      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(comm_nodes[comm][node]);
       int deg = g.nb_neighbors(comm_nodes[comm][node]);
       for (int i=0 ; i<deg ; i++) {
 	int neigh        = *(p.first+i);
@@ -646,7 +646,7 @@ Community::one_level(int W, bool level0) {
 		      if(best_comm!=node_comm){
 		        //add neighbors to the queue, whose community != best_community
 		        //assumption: the neighbor oder here is the same as what appears in neigh_pos[] array
-			pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+			pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
 			unsigned int deg = g.nb_neighbors(node);
 			for(unsigned int i=0; i<deg ; i++){
 			      int neigh = *(p.first+i);
