+++ CE_Louvain_V5_1/community.cpp
@@ -95,7 +95,7 @@ Community::init_partition(char * filename) {
       unsigned int i=0;
       for ( i=0 ; i<neigh_last ; i++) {
 	unsigned int best_comm     = neigh_pos[i];
-	double best_nblinks  = neigh_weight[neigh_pos[i]];
+	float best_nblinks  = neigh_weight[neigh_pos[i]];
 	if (best_comm==comm) {
 	  insert(node, best_comm, best_nblinks);
 	  break;
@@ -153,7 +153,7 @@ Community::neigh_comm(unsigned int node) {
     neigh_weight[neigh_pos[i]]=-1;
   neigh_last=0;
 
-  pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+  pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
 
   unsigned int deg = g.nb_neighbors(node);
 
@@ -194,7 +194,7 @@ Community::miniBatch_assign(long W){
         	    diff = deg - max_deg;
          	    max_deg = deg;
       		}
-		long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(double)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
+		long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(float)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
       		if(sum + tmp + diff *(long)(sizeof(double)+sizeof(unsigned int)) >= W)
         	   break;
       		sum += tmp + diff * (sizeof(double)+sizeof(unsigned int));
@@ -235,7 +235,7 @@ Community::partition2graph() {
 
 
   for (int i=0 ; i<size ; i++) {
-    pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(i);
+    pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(i);
 
     int deg = g.nb_neighbors(i);
     for (int j=0 ; j<deg ; j++) {
@@ -288,12 +288,12 @@ Community::partition2graph_binary() {
 
   int comm_deg = comm_nodes.size();
   for (int comm=0 ; comm<comm_deg ; comm++) {
-    map<int,double> m;
-    map<int,double>::iterator it;
+    map<int,float> m;
+    map<int,float>::iterator it;
 
     int comm_size = comm_nodes[comm].size();
     for (int node=0 ; node<comm_size ; node++) {
-      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(comm_nodes[comm][node]);
+      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(comm_nodes[comm][node]);
       int deg = g.nb_neighbors(comm_nodes[comm][node]);
       for (int i=0 ; i<deg ; i++) {
 	int neigh        = *(p.first+i);
@@ -394,7 +394,7 @@ Community::one_level(int W) {
 		      if(best_comm!=node_comm){
 		        //add neighbors to the queue, whose community != best_community
 		        //assumption: the neighbor oder here is the same as what appears in neigh_pos[] array
-		        pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+		        pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
 		        unsigned int deg = g.nb_neighbors(node);
 		        for(unsigned int i=0; i<deg ; i++){
 		           int neigh = *(p.first+i);
