+++ CE_Louvain_V5_StructuralPrune/community.cpp
@@ -1,4 +1,3 @@
-
 // File: community.h
 // -- community detection source file
 //-----------------------------------------------------------------------------
@@ -134,7 +133,7 @@ Community::prune_graph(int T){
          continue;
       if(index != 0)
          g2.degrees[index] = g2.degrees[index-1]; //initialization
-      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
       for (int i=0 ; i<deg ; i++) {
            int neigh = *(p.first+i);  
            double weight = (g.weights.size()==0)?1.:*(p.second+i);
@@ -160,7 +159,7 @@ Community::prune_graph(int T){
       if(deg > T)
          continue;
       g2.degrees[index] = g2.degrees[index-1] + deg; 
-      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
       for (int i=0 ; i<deg ; i++) {
            int neigh = renumber[*(p.first+i)];
            g2.links[link_index] = neigh;
@@ -205,7 +204,7 @@ Community::init_partition(char * filename) {
       unsigned int i=0;
       for ( i=0 ; i<neigh_last ; i++) {
 	unsigned int best_comm     = neigh_pos[i];
-	double best_nblinks  = neigh_weight[neigh_pos[i]];
+	float best_nblinks  = neigh_weight[neigh_pos[i]];
 	if (best_comm==comm) {
 	  insert(node, best_comm, best_nblinks);
 	  break;
@@ -244,7 +243,7 @@ Community::neigh_comm(unsigned int node) {
     neigh_weight[neigh_pos[i]]=-1;
   neigh_last=0;
 
-  pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+  pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
 
   unsigned int deg = g.nb_neighbors(node);
 
@@ -277,7 +276,7 @@ Community::neigh_comm_low(unsigned int node) {
   int max_inDeg = 0; 
   unsigned int ind = 0; 
 
-  pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+  pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
 
   unsigned int deg = g.nb_neighbors(node);
 
@@ -322,7 +321,7 @@ Community::miniBatch_assign(long W){
         	    diff = deg - max_deg;
          	    max_deg = deg;
       		}
-		long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(double)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
+		long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(float)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
       		if(sum + tmp + diff *(long)(sizeof(double)+sizeof(unsigned int)) >= W)
         	   break;
       		sum += tmp + diff * (sizeof(double)+sizeof(unsigned int));
@@ -395,7 +394,7 @@ Community::update_sum(int curr, int* max_deg, long sum){
          diff = deg - *max_deg;
          *max_deg = deg; 
        }
-       long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(double)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
+       long tmp = sizeof(int) + sizeof(unsigned long) + deg * (sizeof(unsigned int) + sizeof(float)); //Accounting for n2c, g.degrees[], g.links[] and g.weights[]
        sum += tmp + diff * (sizeof(double)+sizeof(unsigned int));
        return sum; 
 }
@@ -479,7 +478,7 @@ Community::ABFS_preorder(vector<int>& retorder, long W){
 
            //insert the neighbors and remove curr from the queue
            int deg = g.nb_neighbors(curr);
-           pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(curr);
+           pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(curr);
            vector<int> children(deg,-1);
           
            int ch_num = 0; 
@@ -578,7 +577,7 @@ Community::transform(long W){
       int node = order[index];
       int deg = g.nb_neighbors(node);
       g2.degrees[index] = (index==0)?deg:g2.degrees[index-1]+deg;
-      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
       for(int i=0; i<deg; i++){
 	  int neigh = *(p.first+i);
 	  g2.links[link_index] = reverse_map[neigh];
@@ -630,7 +629,7 @@ Community::partition2graph() {
 
 
   for (int i=0 ; i<size ; i++) {
-    pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(i);
+    pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(i);
 
     int deg = g.nb_neighbors(i);
     for (int j=0 ; j<deg ; j++) {
@@ -683,12 +682,12 @@ Community::partition2graph_binary() {
 
   int comm_deg = comm_nodes.size();
   for (int comm=0 ; comm<comm_deg ; comm++) {
-    map<int,double> m;
-    map<int,double>::iterator it;
+    map<int,float> m;
+    map<int,float>::iterator it;
 
     int comm_size = comm_nodes[comm].size();
     for (int node=0 ; node<comm_size ; node++) {
-      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(comm_nodes[comm][node]);
+      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(comm_nodes[comm][node]);
       int deg = g.nb_neighbors(comm_nodes[comm][node]);
       for (int i=0 ; i<deg ; i++) {
 	int neigh        = *(p.first+i);
@@ -735,7 +734,7 @@ Community::partition2graph_binary_l0() {
   vector<vector<pair<int, double> > > comm_lowNeighs(final); //keeping track of low degree neighbors of communities 
   for(int i=boundary; i<size; i++){
       int deg = g.nb_neighbors(i);
-      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(i);
+      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(i);
       for(int j=0; j<deg; j++){
           int neigh = *(p.first+j);
           double neigh_weight = (g.weights.size()==0)?1.:*(p.second+j);
@@ -758,12 +757,12 @@ Community::partition2graph_binary_l0() {
   int comm_deg = comm_nodes.size();
   for (int comm=0 ; comm<comm_deg ; comm++) {
 
-    map<int,double> m;
-    map<int,double>::iterator it;
+    map<int,float> m;
+    map<int,float>::iterator it;
 
     int comm_size = comm_nodes[comm].size();
     for (int node=0 ; node<comm_size ; node++) {
-      pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(comm_nodes[comm][node]);
+      pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(comm_nodes[comm][node]);
       int deg = g.nb_neighbors(comm_nodes[comm][node]);
       for (int i=0 ; i<deg ; i++) {
 	int neigh        = *(p.first+i);
@@ -876,7 +875,7 @@ Community::one_level(int W, bool level0) {
 		      if(best_comm!=node_comm){
 		        //add neighbors to the queue, whose community != best_community
 		        //assumption: the neighbor oder here is the same as what appears in neigh_pos[] array
-		        pair<vector<unsigned int>::iterator, vector<double>::iterator> p = g.neighbors(node);
+		        pair<vector<unsigned int>::iterator, vector<float>::iterator> p = g.neighbors(node);
 		        unsigned int deg = g.nb_neighbors(node);
 		        for(unsigned int i=0; i<deg ; i++){
 		           int neigh = *(p.first+i);
@@ -970,3 +969,4 @@ Community::merge_low(){
   }
 }
 //----------------------------------------------------------------------
+
